+++ Help.Domain.Core/HelpServiceAgg/DTOs/HelpRequest/CreateHelpRequestDTO.cs
@@ -7,8 +7,8 @@ namespace Help.Domain.Core.HelpServiceAgg.DTOs.HelpRequest
         public string Title { get; set; }
         public string Description { get; set; }
         public string ExpirationDate { get; set; }
-        public long CustomerId { get; set; }
-        public long ServiceId { get; set; }
+        public int CustomerId { get; set; }
+        public int ServiceId { get; set; }
         public double ProposedPrice { get; set; }
     }
 }
