+++ Help.Domain.Core/HelpServiceAgg/DTOs/HelpServiceCategory/CreateHelpServiceCategoryDTO.cs
@@ -10,6 +10,6 @@ namespace Help.Domain.Core.HelpServiceAgg.DTOs.HelpServiceCategory
     {
         public string Title { get; set; }
         public string Description { get; set; }
-        public long ParentId { get; set; }
+        public int ParentId { get; set; }
     }
 }
