+++ Help.Domain.Core/HelpServiceAgg/DTOs/Proposal/SearchProposaltDTO.cs
@@ -4,6 +4,6 @@ namespace Help.Domain.Core.HelpServiceAgg.DTOs.Proposal
 {
     public class SearchProposaltDTO
     {
-        public long HelpRequestId { get; set; }
+        public int HelpRequestId { get; set; }
     }
 }
