+++ Help.Domain.Core/HelpServiceAgg/DTOs/Skill/CreateSkillDTO.cs
@@ -7,7 +7,7 @@ namespace Help.Domain.Core.HelpServiceAgg.DTOs.Skill
         public string Title { get; set; }
         public string Description { get; set; }
         public Int16 Level { get; set; }
-        public long ServiceId { get; set; }
-        public long CustomerId { get; set; }
+        public int ServiceId { get; set; }
+        public int CustomerId { get; set; }
     }
 }
