+++ Help.Domain.Core/WalletAgg/DTOs/Wallet/CreateWalletDTO.cs
@@ -4,6 +4,6 @@ namespace Help.Domain.Core.WalletAgg.DTOs.Wallet
 {
     public class CreateWalletDTO
     {
-        public long CustomerId { get; private set; }
+        public int CustomerId { get; private set; }
     }
 }
